[powerpc] Add hwcap/hwcap2/platform data to TCB.
This patch adds a new feature for powerpc.  In order to get faster
access to the HWCAP/HWCAP2 bits and platform number (i.e. for
implementing __builtin_cpu_is()/__builtin_cpu_supports() in GCC)
without the overhead of reading from the auxiliary vector, we now
reserve space for them in the TCB.  This is an ABI change for
GLIBC 2.23.

A new versioned symbol 'init_hwcapinfo' is available to get the data
from the auxiliary vector and parse it, and store it for later use
in the TLS initialization code.  This function also has control code
to make sure it is run only once per program execution.

A new API is published in ppc.h to get the HWCAP+HWCAP2 and
platform number data from the TCB as well.

2015-08-12  Carlos Eduardo Seo  <cseo@linux.vnet.ibm.com>

	* manual/platform.texi: Added a manual entry for the new API.
	* sysdeps/powerpc/Makefile: Added testcases and new init_hwcapinfo
	function to the Makefile.
	* sysdeps/powerpc/Versions: Added new init_hwcapinfo symbol to
	GLIBC-2.23.
	* sysdeps/powerpc/hwcapinfo.c: New file.
	(init_hwcapinfo): New function to initialize and parse hwcap, hwcap2
	and platform number information.
	* sysdeps/powerpc/hwcapinfo.h: New file.  Creates global variables
	to store HWCAP+HWCAP2, platform number and control data.
	* sysdeps/powerpc/nptl/tcb-offsets.sym: Added new offests
	for HWCAP+HWCAP2 and platform number in the TCB.
	* sysdeps/powerpc/nptl/tls.h: New functionality.  Stores
	the HWCAP, HWCAP2 and platform number in the TCB.
	(dtv): Added new fields for HWCAP+HWCAP2 and platform number.
	(TLS_INIT_TP): Included calls to add the hwcap and
	at_platform values in the TCB in TP initialization.
	(TLS_DEFINE_INIT_TP): Likewise.
	(THREAD_GET_HWCAP): New macro.
	(THREAD_SET_HWCAP): Likewise.
	(THREAD_GET_AT_PLATFORM): Likewise.
	(THREAD_SET_AT_PLATFORM): Likewise.
	* sysdeps/powerpc/sys/platform/ppc.h: Added new functions
	to get the HWCAP+HWCAP2 and platform number values from the TCB.
	(__ppc_get_hwcap): New function.
	(__ppc_get_at_platform): Likewise.
	* sysdeps/powerpc/test-get_hwcap.c: Testcase for this functionality.
	* sysdeps/unix/sysv/linux/powerpc/dl-support.c: New file.
	Adds support for AT_PLATFORM in static glibc.
	* sysdeps/unix/sysv/linux/powerpc/powerpc32/fpu/libc.abilist:
	Included the new init_hwcapinfo symbol in the ABI list for GLIBC 2.23.
	* sysdeps/unix/sysv/linux/powerpc/powerpc64/libc-le.abilist:
	Likewise.
	* sysdeps/unix/sysv/linux/powerpc/powerpc64/libc.abilist:
	Likewise.